@@ -24,16 +24,16 @@
   },
   "dependencies": {
     "@elastic/elasticsearch": "^7.6.1",
-    "@nestjs/bull": "^0.1.0",
-    "@nestjs/common": "^7.0.0",
-    "@nestjs/config": "^0.4.2",
-    "@nestjs/core": "^7.0.0",
+    "@nestjs/bull": "^0.3.1",
+    "@nestjs/common": "^7.6.17",
+    "@nestjs/config": "^0.6.3",
+    "@nestjs/core": "^7.6.17",
     "@nestjs/elasticsearch": "^7.1.0",
-    "@nestjs/jwt": "^7.0.0",
-    "@nestjs/passport": "^7.0.0",
-    "@nestjs/platform-express": "^7.0.0",
-    "@nestjs/schedule": "^0.3.0",
-    "@nestjs/serve-static": "^2.1.0",
+    "@nestjs/jwt": "^7.2.0",
+    "@nestjs/passport": "^7.1.5",
+    "@nestjs/platform-express": "^7.6.17",
+    "@nestjs/schedule": "^0.4.3",
+    "@nestjs/serve-static": "^2.1.4",
     "angular-expressions": "^1.1.2",
     "bcrypt": "^4.0.1",
     "bull": "^3.13.0",
@@ -53,6 +53,7 @@
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^6.5.4",
+    "sitemapper": "^3.1.12",
     "underscore": "^1.10.2",
     "word2pdf-promises": "^1.2.0"
   },
@@ -67,11 +68,11 @@
     "@types/passport-jwt": "^3.0.3",
     "@types/passport-local": "^1.0.33",
     "@types/supertest": "^2.0.8",
-    "@typescript-eslint/eslint-plugin": "^2.23.0",
-    "@typescript-eslint/parser": "^2.23.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.0",
-    "eslint-plugin-import": "^2.20.1",
+    "@typescript-eslint/eslint-plugin": "^4.6.1",
+    "@typescript-eslint/parser": "^4.6.1",
+    "eslint": "^7.12.1",
+    "eslint-config-prettier": "7.1.0",
+    "eslint-plugin-prettier": "^3.1.4",
     "jest": "^25.1.0",
     "prettier": "^1.19.1",
     "supertest": "^4.0.2",
