@@ -5,20 +5,20 @@ module.exports = {
     sourceType: 'module',
   },
   plugins: ['@typescript-eslint/eslint-plugin'],
-  extends: [
-    'plugin:@typescript-eslint/eslint-recommended',
-    'plugin:@typescript-eslint/recommended',
-    'prettier',
-    'prettier/@typescript-eslint',
-  ],
+  // extends: [
+  //   'plugin:@typescript-eslint/recommended',
+  //   'prettier/@typescript-eslint',
+  //   'plugin:prettier/recommended',
+  // ],
   root: true,
   env: {
     node: true,
     jest: true,
   },
+  ignorePatterns: ['.eslintrc.js'],
   rules: {
-    '@typescript-eslint/interface-name-prefix': 'off',
-    '@typescript-eslint/explicit-function-return-type': 'off',
-    '@typescript-eslint/no-explicit-any': 'off',
+    'no-multi-spaces':1,
+    'no-trailing-spaces': 1,
+    'no-unused-vars':1
   },
 };
